Move queryPrometheus to more common base
