+++ server_utils/resources/promql_scalar.json
