+++ server_utils/resources/promql_vector.json
